Changed int to double wherever applicable and necessary. Merge back into algorithm2 is pending results of correctness test of the old EDP algorithm.

+++ src/Approx.java
@@ -27,7 +27,7 @@ public class Approx {
     }
 
     public int ApproximateOptimalTardiness(double epsilon) {
-        int Tmax = this.greedyScheduleFixed.getTardiness();
+        double Tmax = this.greedyScheduleFixed.getTardiness();
         if(Tmax == 0) {
             return 0;
         }
